requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Werkzeug==2.3.7
python-dotenv==1.0.0
google-generativeai==0.3.1
requests==2.31.0
SQLAlchemy==2.0.23
Flask-Login==0.6.3
Flask-WTF==1.2.1
email-validator==2.1.0.post1
cryptography==41.0.5
Pillow==10.1.0
gunicorn==21.2.0
python-dateutil==2.8.2